data-pipeline/requirements.txt
requests
beautifulsoup4
pandas
lxml
pyyaml
python-dateutil
